+++ Cargo.toml
@@ -1,40 +1,15 @@
-[package]
-name = "beam-file"
-version = "0.1.0"
-edition = "2021"
-license = "Apache-2.0"
+[workspace]
+members = ["beam-file", "beam-file-lib", ]
+resolver = "2"
 
-[lib]
-name = "beam_file"
-path = "src/lib.rs"
-
-[[bin]]
-name = "beam-file"
-path = "src/main.rs"
-
-[dependencies]
-axum = { version = "0.8", optional = true}
-axum-extra = { version = "0.12", optional = true, features = ["typed-header"]}
-http-serde = "2.1"
-reqwest = { version = "0.12", features = ["rustls-tls", "stream"], default-features = false }
-clap = { version = "4.5", features = ["derive", "env"] }
+[workspace.dependencies]
+tracing-subscriber = "0.3.0"
+tracing = "0.1.0"
+anyhow = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 tokio = { version = "1", features = ["macros", "rt-multi-thread", "signal", "fs", "io-std"] }
 once_cell = "1"
+reqwest = { version = "0.12", features = ["rustls-tls", "stream"], default-features = false }
 beam-lib = { git = "https://github.com/samply/beam", branch = "fix/relative-urls", features = ["http-util", "sockets"] }
-serde_json = "1"
-serde = { version = "1", features = ["derive"] }
-tokio-util = { version = "0.7", features = ["io"] }
-futures-util = { version = "0.3", default-features = false, features = ["std"] }
-anyhow = "1"
-sync_wrapper = { version = "1", features = ["futures"] }
-tracing-subscriber = "0.3.22"
-tracing = "0.1.44"
-
-[features]
-server = ["dep:axum", "dep:axum-extra"]
-
-[profile.release]
-lto = true
-opt-level = "z"
-codegen-units = 1
-strip = true
+beam-file-lib = { workspace = true }
