+++ beam-file-lib/src/lib.rs
@@ -1,0 +1,18 @@
+pub mod receiver;
+pub mod sender;
+pub mod utils;
+
+pub fn add(left: u64, right: u64) -> u64 {
+    left + right
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn it_works() {
+        let result = add(2, 2);
+        assert_eq!(result, 4);
+    }
+}
