+++ beam-file-lib/src/receiver/default.rs
@@ -1,14 +1,14 @@
-use crate::utils::config::{FileMeta, CLIENT};
 use anyhow::{bail, Context};
 use beam_lib::SocketTask;
 use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
-use reqwest::Url;
+use reqwest::{Client, Url};
 use std::path::Path;
 use std::time::SystemTime;
 use sync_wrapper::SyncStream;
 use tokio::io::AsyncRead;
 use tokio_util::io::ReaderStream;
 use tracing::info;
+use crate::utils::config::FileMeta;
 
 pub async fn save_file(
     dir: &Path,
@@ -53,6 +53,7 @@ pub async fn forward_file(
     socket_task: SocketTask,
     incoming: impl AsyncRead + Unpin + Send + 'static,
     cb: &Url,
+    client: Client,
 ) -> anyhow::Result<()> {
     let FileMeta {
         suggested_name,
@@ -71,7 +72,7 @@ pub async fn forward_file(
             HeaderValue::from_str(&name)?,
         );
     }
-    let res = CLIENT
+    let res = client
         .post(cb.clone())
         .headers(headers)
         .body(reqwest::Body::wrap_stream(SyncStream::new(
