+++ beam-file/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "beam-file"
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+
+[[bin]]
+name = "beam-file"
+path = "src/main.rs"
+
+[dependencies]
+axum = { version = "0.8", optional = true}
+axum-extra = { version = "0.12", optional = true, features = ["typed-header"]}
+reqwest = { workspace = true }
+clap = { version = "4.5", features = ["derive", "env"] }
+tokio = { workspace = true }
+once_cell = { workspace = true }
+beam-lib = { workspace = true }
+serde_json = { workspace = true }
+serde = { workspace = true }
+tokio-util = { version = "0.7", features = ["io"] }
+futures-util = { version = "0.3", default-features = false, features = ["std"] }
+anyhow = {workspace = true}
+tracing-subscriber = { workspace = true }
+tracing = { workspace = true }
+beam-file-lib = { path = "../beam-file-lib" }
+
+[features]
+server = ["dep:axum", "dep:axum-extra"]
+
+[profile.release]
+lto = true
+opt-level = "z"
+codegen-units = 1
+strip = true
