+++ beam-file/Dockerfile
