--- src/lib.rs
@@ -1,3 +1,0 @@
-pub mod receiver;
-pub mod sender;
-pub mod utils;
